+++ src/main/java/com/handlers/ContactsHandler.java
@@ -1,15 +1,15 @@
 package com.handlers;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 
+import com.dao.AuthorizationDao;
 import com.dao.ContactDao;
 import com.exceptions.DaoException;
 import com.filters.AuthFilter;
 import com.models.Contact;
+import com.models.ContactEmail;
 import com.models.User;
-import com.utils.ContactsUtil;
 import com.utils.ExceptionHandlerUtil;
 import com.utils.PathParamUtil;
 
@@ -18,17 +18,22 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
 public class ContactsHandler {
-    public static void returnSpecificContact(HttpServletRequest request, HttpServletResponse response, int contactId)
+	public static void returnSpecificContact(HttpServletRequest request, HttpServletResponse response, int contactId)
             throws IOException, ServletException {
         try {
             User user = (User) AuthFilter.getCurrentUser();
-            Contact contact = ContactDao.getContactByContactId(contactId, user.getUserId());
-            if (contact != null) {
-                request.setAttribute("contact", contact);
-                request.getRequestDispatcher("/contactView.jsp").forward(request, response);
-            } else {
-                ExceptionHandlerUtil.logAndForwardClientException(request, response, "Contact not found.", null, "/error.jsp", ContactsHandler.class);
+            int userId = user.getUserId();
+            Contact contact = ContactDao.getContactByContactId(contactId, userId);
+            if(contact == null) {
+            	ExceptionHandlerUtil.logAndForwardClientException(request, response, "Contact not found.", null, "/error.jsp", ContactsHandler.class);
+            	return;
             }
+
+            List<ContactEmail> contactEmails = ContactDao.getContactEmailsByContactId(contactId, userId);
+            contact.setEmails(contactEmails);
+
+            request.setAttribute("contact", contact);
+            request.getRequestDispatcher("/contactView.jsp").forward(request, response);
         } catch (DaoException e) {
             ExceptionHandlerUtil.logAndForwardServerException(request, response, e, ContactsHandler.class);
         }
@@ -48,6 +53,7 @@ public class ContactsHandler {
             int limit = 11;
             int offset = pageNumber*10;
             List<Contact> contacts = ContactDao.getContactsByUserId(user.getUserId(),limit, offset);
+            System.out.println(contacts);
             boolean isLastPage = isLastPage(contacts, limit);
             List<Contact> contactsToDisplay = contacts;
             if(!isLastPage) {
@@ -77,11 +83,11 @@ public class ContactsHandler {
             String aliasName = request.getParameter("alias_fnd_name");
             String phone = request.getParameter("phone");
             String address = request.getParameter("address");
-            int userId = Integer.parseInt(request.getParameter("user_id"));
             int isArchived = Integer.parseInt(request.getParameter("isArchived"));
             int isFavorite = Integer.parseInt(request.getParameter("isFavorite"));
+            int userId = AuthFilter.getCurrentUser().getUserId();
 
-            boolean isSuccess = ContactDao.addContact(friendEmail, aliasName, phone, address, userId, isArchived, isFavorite);
+            boolean isSuccess = ContactDao.addContact(friendEmail, aliasName, phone, address, userId, null, isArchived, isFavorite);
             if (isSuccess) {
                 response.sendRedirect("/contacts");
             } else {
@@ -96,18 +102,22 @@ public class ContactsHandler {
             throws ServletException, IOException {
         try {
             int contactId = Integer.parseInt(PathParamUtil.getSingleParam(request.getPathInfo()));
-            String friendEmail = request.getParameter("friend_email");
             String aliasName = request.getParameter("alias_fnd_name");
-            String phone = request.getParameter("phone");
             String address = request.getParameter("address");
             int isArchived = Integer.parseInt(request.getParameter("isArchived"));
             int isFavorite = Integer.parseInt(request.getParameter("isFavorite"));
+            int userId = AuthFilter.getCurrentUser().getUserId();
 
-            Contact contact = new Contact(contactId, aliasName, friendEmail, phone, address, isArchived, isFavorite);
-            boolean isSuccess = ContactDao.updateContact(contact);
+			boolean isAuthorized = AuthorizationDao.isUserAuthorizedForCrudOnContacts(userId, contactId);
+			if(!isAuthorized) {
+			    ExceptionHandlerUtil.logAndForwardClientException(request, response, "You cannot do that", null, "/error.jsp", ContactsHandler.class);
+			    return;
+			}
+
+            boolean isSuccess = ContactDao.updateContact(contactId, aliasName, address, isArchived, isFavorite);
 
             if (isSuccess) {
-                response.sendRedirect("/contacts");
+                response.sendRedirect("/contacts/"+contactId);
             } else {
                 ExceptionHandlerUtil.logAndForwardClientException(request, response, "Failed to edit contact.", null, "/contacts", ContactsHandler.class);
             }
@@ -120,6 +130,13 @@ public class ContactsHandler {
             throws IOException, ServletException {
         try {
             int contactId = Integer.parseInt(PathParamUtil.getSingleParam(request.getPathInfo()));
+            int userId = AuthFilter.getCurrentUser().getUserId();
+
+			boolean isAuthorized = AuthorizationDao.isUserAuthorizedForCrudOnContacts(userId, contactId);
+			if(!isAuthorized) {
+			    ExceptionHandlerUtil.logAndForwardClientException(request, response, "You cannot do that", null, "/error.jsp", ContactsHandler.class);
+			    return;
+			}
             boolean isSuccess = ContactDao.deleteContact(contactId);
 
             if (isSuccess) {
@@ -133,4 +150,106 @@ public class ContactsHandler {
             ExceptionHandlerUtil.logAndForwardServerException(request, response, e, ContactsHandler.class);
         }
     }
+
+	public static void handleAddContactPhone(HttpServletRequest request, HttpServletResponse response) throws IOException {
+        try {
+            int contactId = Integer.parseInt(PathParamUtil.getSingleParam(request.getPathInfo()));
+            String label = request.getParameter("label");
+            String phoneNumber = request.getParameter("phoneNumber");
+            int userId = AuthFilter.getCurrentUser().getUserId();
+
+			boolean isAuthorized = AuthorizationDao.isUserAuthorizedForCrudOnContacts(userId, contactId);
+			if(!isAuthorized) {
+			    ExceptionHandlerUtil.logAndForwardClientException(request, response, "You cannot do that", null, "/error.jsp", ContactsHandler.class);
+			    return;
+			}
+
+            boolean isSuccess = ContactDao.addContactPhoneNumber(contactId, label, phoneNumber);
+            if (isSuccess) {
+                response.sendRedirect("/contacts/"+contactId);
+            } else {
+                ExceptionHandlerUtil.logAndForwardClientException(request, response, "Failed to add contact.", null, "/contacts", ContactsHandler.class);
+            }
+        }catch (DaoException e) {
+            ExceptionHandlerUtil.logAndForwardServerException(request, response, e, ContactsHandler.class);
+        }
+	}
+	
+	public static void handleAddContactEmail(HttpServletRequest request, HttpServletResponse response) throws IOException {
+        try {
+            int contactId = Integer.parseInt(PathParamUtil.getSingleParam(request.getPathInfo()));
+            String label = request.getParameter("label");
+            String email = request.getParameter("email");
+            int userId = AuthFilter.getCurrentUser().getUserId();
+
+			boolean isAuthorized = AuthorizationDao.isUserAuthorizedForCrudOnContacts(userId, contactId);
+			if(!isAuthorized) {
+			    ExceptionHandlerUtil.logAndForwardClientException(request, response, "You cannot do that", null, "/error.jsp", ContactsHandler.class);
+			    return;
+			}
+
+            boolean isSuccess = ContactDao.addContactEmail(contactId, label, email);
+            if (isSuccess) {
+                response.sendRedirect("/contacts/"+contactId);
+            } else {
+                ExceptionHandlerUtil.logAndForwardClientException(request, response, "Failed to add contact.", null, "/contacts", ContactsHandler.class);
+            }
+        }catch (DaoException e) {
+            ExceptionHandlerUtil.logAndForwardServerException(request, response, e, ContactsHandler.class);
+        }
+	}
+
+	public static void handleDeleteContactEmail(HttpServletRequest request, HttpServletResponse response) throws IOException {
+        try {
+			String[] pathParams = PathParamUtil.getMultipleParams(request.getPathInfo());
+			int contactId = Integer.parseInt(pathParams[0]);
+			int contactEmailId = Integer.parseInt(pathParams[1]);
+			int userId = AuthFilter.getCurrentUser().getUserId();
+
+			boolean isAuthorized = AuthorizationDao.isUserAuthorizedToDeleteContactEmail(userId, contactEmailId);
+			if(!isAuthorized) {
+			    ExceptionHandlerUtil.logAndForwardClientException(request, response, "You cannot do that", null, "/error.jsp", ContactsHandler.class);
+			    return;
+			}
+			
+			boolean isSuccess = ContactDao.deleteContactEmail(contactEmailId);
+			if (isSuccess) {
+			    response.sendRedirect("/contacts/"+contactId);
+			} else {
+			    ExceptionHandlerUtil.logAndForwardClientException(request, response, "Failed to delete contact email", null, "/contacts", ContactsHandler.class);
+			}
+		} catch (NumberFormatException e) {
+            ExceptionHandlerUtil.logAndForwardClientException(request, response, "Invalid contact email ID format.", e, "/error.jsp", ContactsHandler.class);
+		} catch (DaoException e) {
+            ExceptionHandlerUtil.logAndForwardServerException(request, response, e, ContactsHandler.class);
+		}
+		
+	}
+
+	public static void handleDeleteContactPhone(HttpServletRequest request, HttpServletResponse response) throws IOException {
+		try {
+			String[] pathParams = PathParamUtil.getMultipleParams(request.getPathInfo());
+			int contactId = Integer.parseInt(pathParams[0]);
+			int contactPhoneNumberId = Integer.parseInt(pathParams[1]);
+			int userId = AuthFilter.getCurrentUser().getUserId();
+			
+			boolean isAuthorized = AuthorizationDao.isUserAuthorizedToDeleteContactNumber(userId, contactPhoneNumberId);
+			if(!isAuthorized) {
+			    ExceptionHandlerUtil.logAndForwardClientException(request, response, "You cannot do that", null, "/error.jsp", ContactsHandler.class);
+			    return;
+			}
+			
+			boolean isSuccess = ContactDao.deleteContactPhone(contactPhoneNumberId);
+			if (isSuccess) {
+			    response.sendRedirect("/contacts/"+contactId);
+			} else {
+			    ExceptionHandlerUtil.logAndForwardClientException(request, response, "Failed to delete contact phone", null, "/contacts", ContactsHandler.class);
+			}
+		} catch (NumberFormatException e) {
+            ExceptionHandlerUtil.logAndForwardClientException(request, response, "Invalid contact phone ID format.", e, "/error.jsp", ContactsHandler.class);
+		} catch (DaoException e) {
+            ExceptionHandlerUtil.logAndForwardServerException(request, response, e, ContactsHandler.class);
+		}
+		
+	}
 }
